requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-jazzmin==3.0.0
django-rest-authtoken==2.1.4
djangorestframework==3.15.1
gunicorn==22.0.0
mypy==1.10.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
